feat: rename Container.tsx to container.tsx
